feat(compose): Implement shouldResetTimerOnRetrigger config
When the flag is on we need to kill any existing timer and start a new one.
When the flag is off we just do nothing if there is a timer, otherwise
start a timer. Added a whole bunch of test cases around it. Existing test
case basically already are testing the true case in many forms.

Also fixed a typo in the README for the flag