fix(compose): Remove allowRefire configuration
The `allowRefire` config isn't needed because we'll allow a composite event
to refire w/o a cancel event happening. This made the config superfluous.
So removed it from the README, test & code.

Also did some restructuring of the code such that `false` explicitly has
to be passed in for `shouldResetTimerOnRetrigger` instead of any falsey
value. Also setting up the code for the `beforeCallback` implementation.