build(rollup): Use rollup to build target formats
Added a new Rollup configuration file to manage building the source code into target module formats: ES, CommonJS & UMD. The `package.json` points to the different target formats. As part of the work, decided to also refactor `compose` to not have any third-party dependencies to lighten the bundle size. Also added tests to ensure that we're not passing along the composite event to the underlying component.