ci: Try using deploy script again
With 34bebb91c5c4b659846c125d935d8545693024cf I moved the release script into `after_success` which worked and released v1! But ideally the release script will be in the deploy script so that we can return to running tests on multiple node versions. Trying to move it back to the deploy script, but keep the `yarn global add` in `after_success`.

Also:
- Adding dummy CHANGELOG.md for when the lib is released. It points to releases Github page
- Updating README to list ESM modules first