build: Use Rollup to generate UMD bundles (#5)
In addition to having lib formats (ESM, CJS & UMD), we should also have a bundled dist in UMD for those clients that basically can't do anything. The dist can be used directly in a `<script>` or used in a legacy module system like requireJS. There's a minified production bundle as well as an uniminified development bundle. Both have sourcemaps.

Basically had to duplicate all of the configuration for both gulp & rollup, but with shared objects things remained DRY.

Also excluded the `src/` directory